requirements.txt
# Core dependencies
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.9.0
pyyaml>=6.0.1
numpy>=1.26.4

# Optional dependencies for advanced features
# scipy==1.12.0  # Commented out - requires Fortran compiler

# Development dependencies (optional)
# pytest==8.1.1
# hypothesis==6.99.13

